int seed zu long seed angepasst

+++ src/X3UR.Domain/Interfaces/IRandomProvider.cs
@@ -4,5 +4,5 @@ public interface IRandomProvider {
     int Next(int maxValue);
     int Next(int minValue, int maxValue);
     double NextDouble();
-    void Reseed(int seed);
+    void Reseed(long seed);
 }
